Super Nice Newtorking Stack
This is mainly extracted out, and generalized version, of sipstack.io
where I started to build a somewhat general purpose networking stack on
top of Netty. I wanted a simple and quick way to get started to build
any type of networking app on top of Netty and this is the first version
(albeit, one could argue that sipstack.io was first).

Also have the intention of moving over the "codec" parts from pkts.io
over to this netwokring stack and then pkts.io will have to reference
this one. Time will tell how it goes...

Documentation: for now, look at the examples directory and unit tests...