Turned all into DiameterEvent instead.
In order to be able to facilitate different types of events, and not
just request/answers, we had to start using "events" instead where the
DiameterMessageEvent is the most common one. However, an application may
want to know about "peer events" as well and those will of course have
to be presented differently than a DiameterMessage. Hence, events...