Allowing to specify callbacks on Transactions
Extending Transactions so that a user can register callbacks for e.g. a
re-transmission or, in general, for when an answer comes within a Transaction
to process it within that context.

Updating the PeerFSM to actually pay attentions to write, which is
needed to handle these "user" transactions. Internally, we'll always
create a transaction and then it is up to the user if they care or not.

Still work to be done...