Introduced PdnSessionContext, removed PdnSession for now.
Using PdnSessionContext for the CreateSessionRequest and
CreateSessionResponse in order to get easy access to all the various
pieces of information needed for managing this session. It is really
just a simple wrapper and is intended to stay that way. If need be,
we'll add a stack managed PdnSession, kind of like a Transaction but
higher level, that will take care of all aspects of manging the session.
Working on that API so we'll see...