Better UDP connection management + allow app to save connection.
UDP is a bit tricky since it is not a connection oriented protocol but
the Snice Networking stack is trying to obtain the same level of
abstract notion of a Connection (including its lifecycle management)
irrespective whether it is UDP or e.g. TCP. This branch is focusing on
cleaning that up.

Also, as an application you may want to save a connection, which for
user initiated outbound connections is done through the CompletionStage
that is returned when asking to connect to a remote endpoint. However,
for incoming connections, there was no good way to get hold of that
connection. Therefore, there is now a new "save" action on the
ConnectionContext builder that allows the app to save the connection
away.

You should not, as the application writer, save the Connection object
given to your app upon the invocation of an incoming message (or event)
since that one isn't "real" but rather a buffered one. If you were to
save this connection, nothing would happen since no one is "monitoring"
that buffered connection after the invocation. Hence, use the one in the
"save" method.