Adding DataTunnel as a convenient way for applications to deal with user data transactions.
If you build an applicaiton representing the UE (user equipment, such as
a modem, your phone etc) you want to send data across a GTP-U tunnel.
That data may be of transactional nature (e.g. a DNS Request expects a
DNS Response, which is matched to the request through a transaction
identifier) and as the UE, you do want to match incoming
request/responses to each other. The network components, such as a SGW,
PGW do not care at all about this data but in order for an application
representing that UE to not have to re-invent this over and over, a new
DataTunnel has been created to just assist for these use cases.

Again, if you're building a network component (SGW, PGW), you do not
care about this. If you're building a end-to-end tester (which then is
pretending to be a UE), you do care.