Added GtpUserTunnel.createBearer
Previously you had to do EpsBearer.create(tunnel, ctx, port) yourself
but this should really be "hidden" and provided to the user in the
context of the established GTP-U tunnel, which it is now. Not much of a
change but also easier to find the operations that the GTP application developer
may want to use.